Update complexinsulincalc.py
Streamlined the code and also added in a question asking if a dose of insulin has been given in the last 90 minutes. If the answer is 'Yes' the script stops and give a dose based upon the carbs x ratio only. If the answer is 'No', the script asks for the Blood Sugar level and then will add in a correctional dose if required.